fix: destroy and create dummy player manually upon every campaign load
It is important to create it only after FirstWorldInit because that is when everything has been properly set up and otherwise the player seems to be in an incompletely initialized state causing many errors and crashes. Similarly it is important to destroy it and recreate it upon every game load otherwise similar errors occur as it goes in a weird state. Examples of weirdness include not properly updating properties in skill container update e.g. when equipping a weapon the damage from the weapon is not added.