fix: hiding damage numbers for real actor skills without items
E.g. skills present on beasts such as ghoul_claws on ghouls. Now we only hide the damage numbers if the skill is present on the dummy player which is an indication that the skill is a nested tooltip in a general sense and not a nested tooltip of a specific skill on a real actor.