fix(vanilla): unequip bagged items upon calling item container unequip
Bug report: https://steamcommunity.com/app/365360/discussions/1/684112192552961717/

This fix is meant to be included in MSU. But we temporarily include it here while MSU is waiting to update, as it is necessary for the proper functioning of the Dummy Player equipping/unequipping process during nested tooltips of skills related to items.

Also includes fix for MSU not calling its skill_container.onUnequip function from item_container.removeFromBag and removeFromBagSlot functions.